test: add temp e2e tests
These are not permanent tests, the script is very very basic. Its sole
purpose is to allow for a chainsaw refactor. These diffs will allow for
extensive changes to be made with assurance that the output is still
correct. Once the refactor is done there should be sufficient native
testing that this e2e suite will not need expansion.