fix(hunt): catch the rest of the builtin detections
When the option to ignore builtins was added grouped builtins were
missed. Now adding these is not really possible as it would be a
breaking change. So for now we create a virtual event with the data
incorrectly typed. This will probably not go live but allows for easy
discussion.