feat: use iterator wrappers to remove unwanted collects
Due to laziness and a huge unwillingness to fight the type checker we
used to just call collect and then iter again, but this means that we
run through the whole array and have additional allocations. Basically
super inefficient. We now use an optional iterator to wrap and remove
the need to do the above.