fix: make hunt command line parsing better
Now when `-r` or `-s` are supplied the non labeled rules option will be
ignored this will prevent the issue of the first path argument then
being incorrectly handled as a rules path. Annoyingly there is no
correct answer for how we should handle this but this default behaviour
seems more reasonable.

Fixes #101