fix: reduce ram usage at the cost of additional cpu cycles
Value from serde_json eats up a huge amount of RAM, thus we need to
store it in a more space efficient way, but this means we have to do
additional serialisation/deserialisation pairs. The current changes
should greatly reduce the RAM but we are not being as CPU efficient as
we could here. Firstly bincode will be much faster, and secondly we
deserialise the same thing multiple times during print which could be
cached. This CPU optimisations can come later.