tweak: futher optimisations by changing the hash algorithm
From benchmarking it was clear that most of the time when Chainsaw is
hunting is spent on accessing values from objects (hashmap lookups). The
preprocssing optimisation was a good step forwards but only had a
noticible impact when single thread or on very large rule sets. It was
still apparent that the lookups were the true bottleneck. Luckily there
is no need for a cryptographically safe hashing algorithm. So swapping
to FxHashMap makes sense as that is what it is for. This optimisation as
noted in the code has not been applied to search but when it is the same
improvement will be seen when using search.