tweak: move multithreading up onto the documents
Due to what is probably laziness multithreading was only shoved on the
rules which after the optimisations was actually now sometimes slower
than single threading. It was time to do the optimisation that was meant
to be done a while ago and process the documents in parallel. That has
been done and now chainsaw is even faster, result...