feat: output jsonl in a stream like fashion for hunting
When hunting all formats result in Chainsaw storing all detections until
all files are processed as this is required for certain output formats.
The caveat is that Chainsaw will start to eat RAM... Therefore for
formats where this sort of collection does not make sense it can be
skipped. This is now the case for JSONL, so if a huge number of files
are being processed then JSONL is probably the output to use for now.