feat: add cache to disk support for jsonl
Due to the design of chainsaw and the need to aggregate over a file, RAM
usage can grow erratically depending on the number of detections within
a file. When it comes to streamable formats like JSONL we can make use
of the disk to cache the detections therefore saving RAM at the cost of
performance. This commit enables this approach.

Fixes #102