fix: prevent deadlock from nested rayon
When calling rayon parallel iterators in a nested fashion where a mutex
can be inbetween will cause a deadlock when threads steal work. This is
hard to trigger but was happening, so we disable the multithreading in
the evtx library as we saturate by processing documents in parallel
rather then parsing them as quickly as possible. From initial testing
these seems to have negligible impact on performance.